contrib/vscode/: debugging with VS Code and gdb
Remove "externalConsole" line in contrib/vscode/init.sh because it
seems to not work for everyone, and after a discussion with Matthieu
Moy and Derrick Stolee, we agreed that it better to let the user choose
what does with this line.

Add a useful links in contrib/vscode/README.md to help the user to
configure his VS Code and how to use the debugging feature.

Add a mention to the README in Documentation/MyFirstContribution.txt
and a part "To convince a newcomer that VS code can help him".

Signed-off-by: COGONI Guillaume <cogoni.guillaume@gmail.com>
Co-authored-by: BRESSAT Jonathan <git.jonathan.bressat@gmail.com>
Helped-by: Matthieu Moy <Matthieu.Moy@univ-lyon1.fr>
Helped-by: Derrick Stolee <derrickstolee@github.com>